refactor widget tasks to make them work w jupytext
